multi_steps/requirement.txt
torch==1.10.0
opencv
torchvision==0.11.1
Pillow==8.4.0
numpy==1.21.4
scipy==1.7.3
scikit-learn==1.0.1
matplotlib==3.4.3
tqdm==4.62.3
gdal
cuda==10.2
